// app/utils/ui-types.ts
export type Theme = "system" | "light" | "dark";

export type Size = "xsmall" | "small" | "medium" | "large" | "xlarge";

export type DialogSize = "xsmall" | "small" | "medium" | "large";

export type TypographyVariant =
  | "h1"
  | "h2"
  | "h3"
  | "h4"
  | "h5"
  | "h6"
  | "p"
  | "span";

export type Order = "ASC" | "DESC";

export type SelectLabelOptions = { label: string; value: string | number }[];